Move types to System.Runtime.InteropServices.Marshalling
